+++ src/test/php/driver/DummyImageDriverTestCase.php
@@ -5,12 +5,12 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @package  net\stubbles\img
+ * @package  stubbles\img
  */
-namespace net\stubbles\img\driver;
+namespace stubbles\img\driver;
 use stubbles\lang\Rootpath;
 /**
- * Test for net\stubbles\img\driver\DummyImageDriver.
+ * Test for stubbles\img\driver\DummyImageDriver.
  *
  * @group  img
  * @group  driver