+++ src/test/php/response/DefaultImageResponseTestCase.php
@@ -5,14 +5,14 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  *
- * @package  net\stubbles\img
+ * @package  stubbles\img
  */
-namespace net\stubbles\img\response;
-use net\stubbles\img\Image;
-use net\stubbles\img\ImageType;
+namespace stubbles\img\response;
+use stubbles\img\Image;
+use stubbles\img\ImageType;
 use stubbles\lang\Rootpath;
 /**
- * Test for net\stubbles\img\response\DefaultImageResponse.
+ * Test for stubbles\img\response\DefaultImageResponse.
  */
 class DefaultImageResponseTestCase extends \PHPUnit_Framework_TestCase
 {
@@ -47,7 +47,7 @@ class DefaultImageResponseTestCase extends \PHPUnit_Framework_TestCase
     public function setUp()
     {
         ImageType::$DUMMY->value()->reset();
-        $this->defaultImageResponse = $this->getMockBuilder('net\stubbles\img\response\DefaultImageResponse')
+        $this->defaultImageResponse = $this->getMockBuilder('stubbles\img\response\DefaultImageResponse')
                                            ->setMethods(array('header', 'sendBody'))
                                            ->getMock();
         $rootpath                   = new Rootpath();